remove pocket core
