+++ libs/extensions/llamaindex/hyperpocket_llamaindex/pocket_llamaindex.py
@@ -15,7 +15,7 @@ class PocketLlamaindex(Pocket):
         if use_profile is not None:
             self.use_profile = use_profile
 
-        tools = [self.get_tool(pk) for pk in self.core.tools.values()]
+        tools = [self.get_tool(pk) for pk in self.tools.values()]
         return tools
 
     def get_tool(self, pocket_tool: Tool) -> BaseTool:
